backend/requirements.txt
Flask==2.0.2
requests==2.27.0
boto3==1.20.49
urllib3==1.26.8
beautifulsoup4 == 4.10.0
pymysql==1.0.2